requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# DS4A_Team_14_Capstone/extraction_demo.py: 4
# DS4A_Team_14_Capstone/extraction_full.py: 8
# DS4A_Team_14_Capstone/extraction_funcs/get_data.py: 12
# DS4A_Team_14_Capstone/transform.py: 6
PyYAML == 6.0

# DS4A_Team_14_Capstone/extraction_demo.py: 5
# DS4A_Team_14_Capstone/extraction_funcs/get_data.py: 11
# DS4A_Team_14_Capstone/infrastructure_funcs/aws_infra_funcs.py: 4
# DS4A_Team_14_Capstone/infrastructure_funcs/redshift_cluster.py: 4
boto3 == 1.21.32

# DS4A_Team_14_Capstone/pair_members.py: 9
# DS4A_Team_14_Capstone/transform.py: 2
numpy == 1.21.5

# DS4A_Team_14_Capstone/extraction_demo.py: 7
# DS4A_Team_14_Capstone/extraction_full.py: 6
# DS4A_Team_14_Capstone/extraction_funcs/extraction_processing.py: 7
# DS4A_Team_14_Capstone/extraction_funcs/get_data.py: 10
# DS4A_Team_14_Capstone/pair_members.py: 8
# DS4A_Team_14_Capstone/transform.py: 1
pandas == 1.4.2

# DS4A_Team_14_Capstone/extraction_demo.py: 6
# DS4A_Team_14_Capstone/extraction_full.py: 5
# DS4A_Team_14_Capstone/extraction_funcs/get_data.py: 9
requests == 2.27.1
